Extract routes into separate chunks
