+++ package.json
@@ -101,7 +101,6 @@
     "./src/lib/storeErrorHandler.js",
     "./src/lib/switchWebmanifest.js",
     "./src/store/index.js",
-    "./src/register-modals.js",
     "./src/ui-common.js",
     "*.css",
     "*.scss",
