+++ src/ui.js
@@ -1,15 +1,16 @@
 import Vue from 'vue';
 import Router from 'vue-router';
 import './ui-common';
-import './register-modals';
+import registerModals from './router/modals';
 import sync from './lib/vuexRouterSync';
 import VeeValidate from './lib/veeValidatePlugin';
-import App from './App.vue';
-import AppDesktop from './AppDesktop.vue';
 import router from './router';
 import store from './store';
 import uiPlugin from './store/plugins/ui';
 import setupAnalytics from './setupAnalytics';
+
+const AppMobile = () => import(/* webpackChunkName: "ui-mobile" */ './App.vue');
+const AppDesktop = () => import(/* webpackChunkName: "ui-desktop" */ './AppDesktop.vue');
 
 Vue.use(Router);
 Vue.use(VeeValidate);
@@ -20,9 +21,10 @@ if (process.env.UNFINISHED_FEATURES) {
 
 sync(store, router);
 uiPlugin(store);
+registerModals();
 
 new Vue({
   store,
   router,
-  render: h => h(process.env.IS_MOBILE_DEVICE ? App : AppDesktop),
+  render: h => h(process.env.IS_MOBILE_DEVICE ? AppMobile : AppDesktop),
 }).$mount('#app');
